fix #413: Remove RemoteExceptions utility
Discussed in previous PR, this was left to allow graceful library
upgrades.
