+++ .palantir/revapi.yml
@@ -13,3 +13,16 @@ acceptedBreaks:
       old: null
       new: "method com.palantir.dialogue.Clients com.palantir.dialogue.ConjureRuntime::clients()"
       justification: "alpha software, runtime not meant for extension"
+  "0.12.2":
+    com.palantir.dialogue:dialogue-apache-hc4-client: []
+    com.palantir.dialogue:dialogue-blocking-channels: []
+    com.palantir.dialogue:dialogue-core: []
+    com.palantir.dialogue:dialogue-httpurlconnection-client: []
+    com.palantir.dialogue:dialogue-java-client: []
+    com.palantir.dialogue:dialogue-okhttp-client: []
+    com.palantir.dialogue:dialogue-serde: []
+    com.palantir.dialogue:dialogue-target:
+    - code: "java.class.removed"
+      old: "class com.palantir.dialogue.RemoteExceptions"
+      new: null
+      justification: "alpha"
